Add Creep and Path classes

+++ src/Drawable.h
@@ -13,8 +13,8 @@ public:
 protected:
 	sf::Sprite getSpriteById(int id, sf::Texture&, int tileSize);
 	int tileSize_;
-	int x_;
-	int y_;
+	float x_;
+	float y_;
 	static int drawnSize_;
 
 };
